# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CATEDRALX - Chalamandra Labs SNC API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  /** @nullable */
  name?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserInfo {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  tier: string;
}

export interface AuthResponse {
  user: UserInfo;
}

/**
 * @nullable
 */
export type MeResponseUser = {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  tier: string;
} | null;

export interface MeResponse {
  /** @nullable */
  user: MeResponseUser;
}

export type VitalsDataBtc = {
  price: number;
  change24h: number;
  volatility: number;
};

export type VitalsDataServer = {
  latencyMs: number;
  uptimeHours: number;
  requestCount: number;
  status: string;
};

/**
 * @nullable
 */
export type VitalsDataSignalsLastSignal = { [key: string]: unknown } | null;

export type VitalsDataSignals = {
  /** @nullable */
  lastSignal?: VitalsDataSignalsLastSignal;
  totalCount: number;
};

export type VitalsDataTrading = {
  tradeCount: number;
  pnlPercent: number;
};

export type VitalsDataUser = {
  tier: string;
  sessionActive: boolean;
  hasStripeSubscription: boolean;
  secureConnection: boolean;
};

export type VitalsDataEntropy = {
  repoCount: number;
  signalsAccumulated: number;
  tradesExecuted: number;
};

export type VitalsDataAffective = {
  valencia: number;
  activacion: number;
  persistencia: number;
};

export interface VitalsData {
  btc: VitalsDataBtc;
  server: VitalsDataServer;
  signals: VitalsDataSignals;
  trading: VitalsDataTrading;
  user: VitalsDataUser;
  entropy: VitalsDataEntropy;
  affective: VitalsDataAffective;
}

export type NervousStateGwtFocusItem = {
  variable: string;
  weight: number;
};

export type NervousStateModuleActivations = { [key: string]: number };

export type NervousStatePulse = {
  dopamine: number;
  adrenaline: number;
  soma: number;
  asymmetry: number;
  liquidityTrap: boolean;
};

export interface NervousState {
  phi: number;
  delta: number;
  gwtFocus: NervousStateGwtFocusItem[];
  gwtBroadcast: string;
  synapticWeights: number[][];
  strategyNames?: string[];
  learningRate: number;
  moduleActivations: NervousStateModuleActivations;
  pulse: NervousStatePulse;
  timestamp: string;
}

export interface AffectiveBody {
  source: string;
}

export interface Signal {
  id?: number;
  strategy: string;
  type: string;
  message: string;
  price: number;
  confidence: number;
  created_at?: string;
}

export interface SignalsResponse {
  signals: Signal[];
  limited: boolean;
  tier: string;
}

export interface GenerateSignalsResponse {
  signals: Signal[];
  count: number;
}

/**
 * @nullable
 */
export type PortfolioResponsePortfolio = { [key: string]: unknown } | null;

export type PortfolioResponseTradesItem = { [key: string]: unknown };

export interface PortfolioResponse {
  /** @nullable */
  portfolio: PortfolioResponsePortfolio;
  trades: PortfolioResponseTradesItem[];
  btcPrice: number;
}

export interface TradeBody {
  action: string;
  amountUsd: number;
  /** @nullable */
  strategy?: string | null;
}

export type TradeResponsePortfolio = { [key: string]: unknown };

export interface TradeResponse {
  success: boolean;
  message: string;
  portfolio?: TradeResponsePortfolio;
}

export type TradeHistoryResponseTradesItem = { [key: string]: unknown };

export interface TradeHistoryResponse {
  trades: TradeHistoryResponseTradesItem[];
}

export interface CommitBody {
  message: string;
}

export interface GitAddBody {
  /** @nullable */
  files?: string | null;
}

export interface GitOperationResponse {
  status: string;
  message: string;
}

export interface Repo {
  id: string;
  nombre: string;
  capa: string;
  peso_w: number;
  descripcion: string;
  url: string;
}

export interface StrategyInvokeResponse {
  status: string;
  message: string;
  id: number;
}

export interface AiGenerateBody {
  lobe: string;
  prompt: string;
}

export interface AiGenerateResponse {
  text: string;
  lobe: string;
  model: string;
}

export type Logout200 = {
  success: boolean;
};

export type RecordAffective200 = {
  success: boolean;
};

export type GetSignalsParams = {
  limit?: number;
};

export type GetGitStatus200 = { [key: string]: unknown };

export type GetGithubRepos200Item = { [key: string]: unknown };