# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CATEDRALX - Chalamandra Labs SNC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /paper-trading/trade:
    post:
      operationId: executeTrade
      tags: [paper-trading]
      summary: Execute a paper trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeBody"
      responses:
        "200":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResponse"
...
  /git/commit:
    post:
      operationId: gitCommit
      tags: [git]
      summary: Commit changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitBody"
      responses:
        "200":
          description: Committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitOperationResponse"
...
  /git/add:
    post:
      operationId: gitAdd
      tags: [git]
      summary: Stage files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitAddBody"
      responses:
        "200":
          description: Files staged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitOperationResponse"
...
  /github/repos/{username}:
    get:
      operationId: getGithubRepos
      tags: [git]
      summary: Get GitHub repos for a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: GitHub repos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
...
  /ai/generate:
    post:
      operationId: aiGenerate
      tags: [ai]
      summary: Generate AI response using Gemini (Soberano only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGenerateResponse"
        "403":
          description: Requires Soberano plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: ["string", "null"]
      required: [email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        tier:
          type: string
      required: [id, email, tier]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
      required: [user]
...
    TradeBody:
      type: object
      properties:
        action:
          type: string
        amountUsd:
          type: number
        strategy:
          type: ["string", "null"]
      required: [action, amountUsd]
    TradeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        portfolio:
          type: object
          additionalProperties: true
      required: [success, message]
...
    CommitBody:
      type: object
      properties:
        message:
          type: string
      required: [message]
    GitAddBody:
      type: object
      properties:
        files:
          type: ["string", "null"]
    GitOperationResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required: [status, message]
...
    AiGenerateBody:
      type: object
      properties:
        lobe:
          type: string
        prompt:
          type: string
      required: [lobe, prompt]
    AiGenerateResponse:
      type: object
      properties:
        text:
          type: string
        lobe:
          type: string
        model:
          type: string
      required: [text, lobe, model]