# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CATEDRALX - Chalamandra Labs SNC API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  email: zod.string(),
  password: zod.string(),
  name: zod.string().nullish(),
});

export const RegisterResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string().nullish(),
    tier: zod.string(),
  }),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string().nullish(),
    tier: zod.string(),
  }),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  user: zod
    .object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string().nullish(),
      tier: zod.string(),
    })
    .nullable(),
});

/**
 * @summary Get system vitals including BTC price, signals, trading stats
 */
export const GetVitalsResponse = zod.object({
  btc: zod.object({
    price: zod.number(),
    change24h: zod.number(),
    volatility: zod.number(),
  }),
  server: zod.object({
    latencyMs: zod.number(),
    uptimeHours: zod.number(),
    requestCount: zod.number(),
    status: zod.string(),
  }),
  signals: zod.object({
    lastSignal: zod.record(zod.string(), zod.unknown()).nullish(),
    totalCount: zod.number(),
  }),
  trading: zod.object({
    tradeCount: zod.number(),
    pnlPercent: zod.number(),
  }),
  user: zod.object({
    tier: zod.string(),
    sessionActive: zod.boolean(),
    hasStripeSubscription: zod.boolean(),
    secureConnection: zod.boolean(),
  }),
  entropy: zod.object({
    repoCount: zod.number(),
    signalsAccumulated: zod.number(),
    tradesExecuted: zod.number(),
  }),
  affective: zod.object({
    valencia: zod.number(),
    activacion: zod.number(),
    persistencia: zod.number(),
  }),
});

/**
 * @summary Get current nervous system state (IIT Phi, GWT, etc.)
 */
export const GetNervousStateResponse = zod.object({
  phi: zod.number(),
  delta: zod.number(),
  gwtFocus: zod.array(
    zod.object({
      variable: zod.string(),
      weight: zod.number(),
    }),
  ),
  gwtBroadcast: zod.string(),
  synapticWeights: zod.array(zod.array(zod.number())),
  strategyNames: zod.array(zod.string()).optional(),
  learningRate: zod.number(),
  moduleActivations: zod.record(zod.string(), zod.number()),
  pulse: zod.object({
    dopamine: zod.number(),
    adrenaline: zod.number(),
    soma: zod.number(),
    asymmetry: zod.number(),
    liquidityTrap: zod.boolean(),
  }),
  timestamp: zod.string(),
});

/**
 * @summary Trigger a nervous system tick (Soberano only)
 */
export const NervousTickResponse = zod.object({
  phi: zod.number(),
  delta: zod.number(),
  gwtFocus: zod.array(
    zod.object({
      variable: zod.string(),
      weight: zod.number(),
    }),
  ),
  gwtBroadcast: zod.string(),
  synapticWeights: zod.array(zod.array(zod.number())),
  strategyNames: zod.array(zod.string()).optional(),
  learningRate: zod.number(),
  moduleActivations: zod.record(zod.string(), zod.number()),
  pulse: zod.object({
    dopamine: zod.number(),
    adrenaline: zod.number(),
    soma: zod.number(),
    asymmetry: zod.number(),
    liquidityTrap: zod.boolean(),
  }),
  timestamp: zod.string(),
});

/**
 * @summary Record an affective state event
 */
export const RecordAffectiveBody = zod.object({
  source: zod.string(),
});

export const RecordAffectiveResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get trading signals
 */
export const GetSignalsQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
});

export const GetSignalsResponse = zod.object({
  signals: zod.array(
    zod.object({
      id: zod.number().optional(),
      strategy: zod.string(),
      type: zod.string(),
      message: zod.string(),
      price: zod.number(),
      confidence: zod.number(),
      created_at: zod.string().optional(),
    }),
  ),
  limited: zod.boolean(),
  tier: zod.string(),
});

/**
 * @summary Manually trigger signal generation
 */
export const GenerateSignalsResponse = zod.object({
  signals: zod.array(
    zod.object({
      id: zod.number().optional(),
      strategy: zod.string(),
      type: zod.string(),
      message: zod.string(),
      price: zod.number(),
      confidence: zod.number(),
      created_at: zod.string().optional(),
    }),
  ),
  count: zod.number(),
});

/**
 * @summary Get current paper trading portfolio
 */
export const GetPortfolioResponse = zod.object({
  portfolio: zod.record(zod.string(), zod.unknown()).nullable(),
  trades: zod.array(zod.record(zod.string(), zod.unknown())),
  btcPrice: zod.number(),
});

/**
 * @summary Execute a paper trade
 */
export const ExecuteTradeBody = zod.object({
  action: zod.string(),
  amountUsd: zod.number(),
  strategy: zod.string().nullish(),
});

export const ExecuteTradeResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  portfolio: zod.record(zod.string(), zod.unknown()).optional(),
});

/**
 * @summary Get trade history
 */
export const GetTradeHistoryResponse = zod.object({
  trades: zod.array(zod.record(zod.string(), zod.unknown())),
});

/**
 * @summary Get git status
 */
export const GetGitStatusResponse = zod.record(zod.string(), zod.unknown());

/**
 * @summary Commit changes
 */
export const GitCommitBody = zod.object({
  message: zod.string(),
});

export const GitCommitResponse = zod.object({
  status: zod.string(),
  message: zod.string(),
});

/**
 * @summary Push to remote
 */
export const GitPushResponse = zod.object({
  status: zod.string(),
  message: zod.string(),
});

/**
 * @summary Stage files
 */
export const GitAddBody = zod.object({
  files: zod.string().nullish(),
});

export const GitAddResponse = zod.object({
  status: zod.string(),
  message: zod.string(),
});

/**
 * @summary Get local repos from chalamandra-labs/repos.json
 */
export const GetLocalReposResponseItem = zod.object({
  id: zod.string(),
  nombre: zod.string(),
  capa: zod.string(),
  peso_w: zod.number(),
  descripcion: zod.string(),
  url: zod.string(),
});
export const GetLocalReposResponse = zod.array(GetLocalReposResponseItem);

/**
 * @summary Get GitHub repos for a user
 */
export const GetGithubReposParams = zod.object({
  username: zod.coerce.string(),
});

export const GetGithubReposResponseItem = zod.record(
  zod.string(),
  zod.unknown(),
);
export const GetGithubReposResponse = zod.array(GetGithubReposResponseItem);

/**
 * @summary Invoke a strategy by ID
 */
export const InvokeStrategyParams = zod.object({
  id: zod.coerce.number(),
});

export const InvokeStrategyResponse = zod.object({
  status: zod.string(),
  message: zod.string(),
  id: zod.number(),
});

/**
 * @summary Generate AI response using Gemini (Soberano only)
 */
export const AiGenerateBody = zod.object({
  lobe: zod.string(),
  prompt: zod.string(),
});

export const AiGenerateResponse = zod.object({
  text: zod.string(),
  lobe: zod.string(),
  model: zod.string(),
});