# lib/api-zod/src/generated/types/gitOperationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CATEDRALX - Chalamandra Labs SNC API
 * OpenAPI spec version: 0.1.0
 */

export interface GitOperationResponse {
  status: string;
  message: string;
}